[DELETE] Remove unused profile components for cleaner codebase
- Deleted CreatedSection, JoinedSection, PetCardSection, ReviewSection, ReviewSubTabs, WritableReviewSection, and WrittenReviewSection components to streamline the profile module.
- Updated ProfileTabContent to reflect the new import paths for the remaining components.
