+++ src/app/profile/_components/JoinedSection/JoinedSection.tsx
