+++ src/app/profile/_components/PetCardSection/PetCardSection.tsx
