+++ src/app/profile/_components/ReviewSection/index.ts
@@ -1,0 +1,3 @@
+export { default as ReviewSection } from "./ReviewSection";
+export { default as WritableReviewSection } from "./WritableReviewSection";
+export { default as WrittenReviewSection } from "./WrittenReviewSection";
